Extract proposal memo contract boundary
